Fixed rounding error in drawing floor
also rewrote the click pixel to unit conversion

+++ src/Floor.cpp
@@ -2,7 +2,7 @@
 #include "Floor.h"
 
 
-int Floor::width, Floor::height;
+float Floor::width, Floor::height;
 
 Floor::Floor(RenderState::StateBlock* stateBlock, float x, float y): x(x), y(y) {
     this->model = Model::create(getMesh());
